fix: 컴포넌트 정의 수정
