+++ src/components/HeaderWithLogo/styles.ts
@@ -1,0 +1,25 @@
+import { css } from 'styled-system/css';
+
+const header = css({
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  height: '4.8rem',
+  padding: '0 1.7rem',
+  borderBottomWidth: '0.1rem',
+  borderBottomColor: 'gray.20',
+});
+
+const logo = css({
+  cursor: 'pointer',
+});
+
+const box = css({
+  width: '3rem',
+  height: '3rem',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+});
+
+export { header, logo, box };
